refactor: script files
