+++ apps/picsa-server/scripts/crop-probability-parser/src/paths.ts
@@ -1,0 +1,7 @@
+import { resolve } from 'path';
+
+export const PATHS = {
+  inputDir: resolve(__dirname, '../', 'input'),
+  outputDir: resolve(__dirname, '../', 'output'),
+  tmpDir: resolve(__dirname, '../', 'tmp'),
+};
